modularize api functions add user api endpoints
